[add] 秒数をいい感じに表示するように
